extended file api

+++ file/src/peers/known_peers.rs
@@ -1,3 +1,5 @@
+use peers::Peer;
+
 use std::fs;
 use std::fs::File;
 use std::io::{BufReader, Read, Write};
@@ -5,12 +7,10 @@ use std::io::{BufReader, Read, Write};
 const BASE_PATH: &str = "./peers";
 const PATH_LAST_PEER: &str = "./peers/last_peer";
 
-/// Contains all information about a peer
+/// Manages the saving and retrieving of known peers
 pub struct KnownPeers {
-    /// Name of the peer
-    pub name: String,
-    /// Socket addr of the peer
-    pub socket: String,
+    /// Instance of peer
+    pub peer: Peer
 }
 
 impl KnownPeers {
@@ -18,8 +18,10 @@ impl KnownPeers {
     ///
     /// Should be called before all other methods
     pub fn init() {
-        fs::create_dir(BASE_PATH).unwrap();
-        File::create(PATH_LAST_PEER).unwrap();
+        if !Path::new(BASE_PATH).exists() {
+            fs::create_dir(BASE_PATH).unwrap();
+            File::create(PATH_LAST_PEER).unwrap();
+        }
     }
 
     /// Removes the peers folder
@@ -37,10 +39,9 @@ impl KnownPeers {
     /// # Return
     ///
     /// Instance of itself
-    pub fn new(name: String, socket: String) -> Self {
+    pub fn new(peer: Peer) -> Self {
         KnownPeers {
-            name: name,
-            socket: socket,
+            peer: peer
         }
     }
 
@@ -51,12 +52,11 @@ impl KnownPeers {
     /// # Returns
     ///
     /// Instance of iteself
-    pub fn save(self) -> Self {
-        let mut file = File::create(format!("peers/{}", self.name)).unwrap();
-        file.write_all(format!("{}\n{}", self.name, self.socket).as_bytes()).unwrap();
+    pub fn save(mut self) -> Self {
+        self.peer = self.peer.save();
 
         let mut file = File::create(PATH_LAST_PEER).unwrap();
-        file.write_all(self.name.as_bytes()).unwrap();
+        file.write_all(self.peer.get_name().as_bytes()).unwrap();
         self
     }
 
@@ -79,13 +79,11 @@ impl KnownPeers {
             let mut lines = content.lines();
 
             KnownPeers {
-                name: String::from(lines.next().unwrap()),
-                socket: String::from(lines.next().unwrap()),
+                peer: Peer::new(String::from(lines.next().unwrap()), String::from(lines.next().unwrap()))
             }
         } else {
             KnownPeers {
-                name: String::from(""),
-                socket: String::from(""),
+                peer: Peer::new(String::from(""), String::from(""))
             }
         }
     }
